notebook-backend/requirements.old.txt
fastapi==0.103.1
psycopg2-binary==2.9.9
uvicorn[standard]==0.23.2
pydantic==1.10.12
websockets>=10.0
langchain==0.0.242
qdrant-client>=1.6.0

# sqlalchemy==2.0.21
langgraph>=0.0.16
langchain-core>=0.1.4
# alembic==1.12.0  # 注释掉旧版本，避免冲突
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
langchain-dashscope>=0.1.8
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# 文档处理
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
chardet==5.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.4.4
html2text==2020.1.16

# PostgreSQL 相关依赖
psycopg2-binary==2.9.9  # 如遇安装问题可先注释掉
alembic==1.13.1
# sqlalchemy==2.0.21

# Celery和Redis依赖
celery>=5.2.0
redis>=4.0.0

langchain-openai>=0.0.1,<0.1.0

email-validator==2.1.0

pydantic-settings==1.99